requirements.txt
Flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.0
simple-websocket==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-nmap==0.7.1
dnspython==2.4.2
urllib3==2.1.0
fpdf==1.7.2
aiohttp==3.9.1
scapy==2.5.0
pytest==7.4.3
python-whois==0.8.0
pyyaml==6.0.1